src/python/requirements.txt
certifi
numpy
pysam
PyVCF
six
sortedcontainers
